refactor: split base filter config for gold and silver
