+++ src/bioetl/domain/filtering/__init__.py
@@ -7,6 +7,7 @@ This package provides:
 All public classes are re-exported from this module for backwards compatibility.
 """
 
+from bioetl.domain.filtering._base_filter_config import BaseFilterConfig
 from bioetl.domain.filtering.column_filter import FilterOperator, GoldColumnFilter
 from bioetl.domain.filtering.gold_config import GoldFilterConfig
 from bioetl.domain.filtering.input_config import FilterColumn, InputFilterConfig
@@ -19,6 +20,7 @@ from bioetl.domain.filtering.range_filter import GoldRangeFilter
 from bioetl.domain.filtering.silver_config import SilverFilterConfig
 
 __all__ = [
+    "BaseFilterConfig",
     "FilterColumn",
     "FilterLoadResult",
     "FilterOperator",
